test de Barrowrepository fait

+++ data/src/main/java/fr/uga/l3miage/library/data/repo/BorrowRepository.java
@@ -61,8 +61,8 @@ public class BorrowRepository implements CRUDRepository<String, Borrow> {
      */
     public int countBorrowedBooksByUser(long userId) {
     // public int countBorrowedBooksByUser(String userId) {
-        int res = entityManager.createQuery("select count(bk) from Borrow b join b.books bk where b.borrower.id = :userId", int.class).setParameter("userId", userId).getSingleResult();
-        return res;
+        long res = entityManager.createQuery("select count(bk) from Borrow b join b.books bk where b.borrower.id = :userId", int.class).setParameter("userId", userId).getSingleResult();
+        return ((int)res);
     }
 
     /**
